Move RecordBuilder and DefinitionResolver into consumer package
